requirements.txt
requests>=2.31.0
PySide6>=6.6
qasync>=0.24.0
PyYAML>=6.0
python-dateutil>=2.9.0
typing-extensions>=4.10.0
lxml>=5.1.0
aiofiles>=23.2.1
